Config.in: Add a menu around BR2_EXTERNAL configuration options
To make it clear to the user that these options are external to Buildroot
and no support is provided by the Buildroot community.

With this, there's no need to recommend people to their options inside
a menu, so remove that from the documentation.

Kconfig nicely allows us to hide the menu completely if BR2_EXTERNAL isn't
used, so make use of that to not confuse people. It would be nice if we
could add some help text to explain the BR2_EXTERNAL stuff as well, but
that isn't supported on menus :/

(cherry picked from commit f4ea4b9824138c7e24c9e4876ff09f14e5d48e87)

Change-Id: I0ed1911b9e7a951a2515b65b16ce8afb72939d99
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>