pkg-autotools: fix patching libtool for version 2.4
If the libtool used by the package is 2.4 (i.e. with no patchlevel), we
end up with a confusing message:

    /bin/sh: line 0: test: -gt: unary operator expected

That's because patchlevel is empty, and thus there is nothing to comapre
to 2, and test complains.

Fake a patchlevel of 0 (which is really what a missing patchlevel means
anyway) in that case.

(cherry picked from commit 383287ce33c7e8780db3b2a3374aa0b2e57af889)

Change-Id: I8258e27db1b602d0c0e9a2361e3e8fd01d5db639
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Peter Korsgaard <jacmet@uclibc.org>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>