pkg: support $(PKG)_OVERRIDED2_SRCDIR not to rsync source
As linux and the package based on autoconf support to compile
not in the source directory, it's not recommended to build the
package within the r-synced directories, which can save disk
space.

$(PKG)_OVERRIED2_SRCDIR is for the purpose to deny the r-sync
and build with the original source. Only the package can be
built independent on the source directory can use the marco.

Change-Id: Iafb6a4eef02e4fcd7f57fa54b1f74f0bfb87516c