kconfig: support to detect conf file location
To provide the possible compatibilities of external files, it'll
be friendly not to write an absolute directory for the file
Config.in.

To support the function, it'll be easy to convert the relative
file name to the absolute one when handling the key word "source",
which is followed by the configuration file. The policy should be:

 - don't handle the absolute file name and the one containing a
   variable with dollar
 - join the file name with the path of current file. detect the
   existence of the file with this and its parents till the file
   is detected

Change-Id: I8bab80cb02c06c5825fae0c65c2ed0788522d96f