automake: create directories for aclocal directories
the directoires are for aclocal commands. once any of them is missed,
aclocal will execute with errors and fail the make processes.

it seems old make (v3.81) has the problem to evalute the command line
before and after the creation of the directory. it's good to create
the directories when starting the build.

Change-Id: Ic3d7a9c8ab2c0a9a3694e3e6b0f4ed07ae6981cf